[update]部屋のアイコンに時間の情報を追加

+++ Assets/Scripts/Title/CreateNewRoomPanel.cs
@@ -123,7 +123,7 @@ namespace AoAndSugi
                     { "PlayerCount", new MaxTeamCount(){ Value = (int)(playerCount.Value) } },
                     { "NpcCount", new MaxTeamCount(){ Value = (int)(npcCount.Value) } },
                     { "BordSize", new BoardSize(){ Value = new int2(){ x = width.Value, y = height.Value }} },
-                    { "MatchTime", new BoardSize(){ Value = (int)(matchTime.Value) } },
+                    { "MatchTime", new BoardSize(){ Value = matchTime.Value } },
                    },
                 CustomRoomPropertiesForLobby = new[] {
                     "DisplayName",
